% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{TileDBArray-pkg}
\alias{TileDBArray-pkg}
\title{The \pkg{TileDBArray} package}
\description{
Implements the TileDB framework as a \linkS4class{DelayedArray} backend,
with read and write functionality for both dense and sparse arrays.
Currently only integer, logical and double-precision values are supported.
}
\author{
Aaron Lun
}
