% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{TileDBArray-globals}
\alias{TileDBArray-globals}
\alias{getTileDBPath}
\alias{setTileDBPath}
\alias{getTileDBSparse}
\alias{setTileDBSparse}
\alias{getTileDBAttr}
\alias{setTileDBAttr}
\alias{getTileDBExtent}
\alias{setTileDBExtent}
\alias{setTileDBContext}
\title{TileDBArray global options}
\usage{
getTileDBPath()

setTileDBPath(path = NULL)

getTileDBSparse()

setTileDBSparse(sparse = NULL)

getTileDBAttr()

setTileDBAttr(attr = NULL)

getTileDBExtent()

setTileDBExtent(extent = NULL)

setTileDBContext(context = NULL)
}
\arguments{
\item{path}{String containing a path to a TileDB backend.}

\item{sparse}{Logical scalar indicating whether sparse backend should be created.}

\item{attr}{String containing the name of a TileDB attribute.}

\item{extent}{Integer scalar specifying the tile extent for all dimensions.
Alternatively, an integer vector of length equal to the number of dimensions,
specifying a different extent for each dimension in the array to be created.}

\item{context}{A TileDB context object, see \code{\link{tiledb_ctx}} for an example.}
}
\value{
All of the getter functions return the current global value,
or a default value if the former is \code{NULL}:
\itemize{
\item \code{path} defaults to a temporary file in \code{\link{tempdir}}.
\item \code{attr} defaults to \code{"x"}.
\item \code{sparse} defaults to \code{FALSE}.
\item \code{extent} defaults to \code{100L}.
\item \code{context} defaults to the value of \code{\link{tiledb_ctx}()}.
}

All setter functions change the global value and return \code{NULL} invisibly.
}
\description{
Global options for (mostly) writing of TileDBArray backends.
}
\seealso{
\code{\link{writeTileDBArray}}, where these functions are most often used.
}
\author{
Aaron Lun
}
