% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TileDBRealizationSink.R
\name{TileDBRealizationSink}
\alias{TileDBRealizationSink}
\alias{writeTileDBArray}
\alias{coerce,TileDBRealizationSink,TileDBMatrix-method}
\alias{coerce,TileDBRealizationSink,TileDBArray-method}
\alias{coerce,TileDBRealizationSink,DelayedArray-method}
\alias{coerce,ANY,TileDBArray-method}
\alias{coerce,ANY,TileDBMatrix-method}
\alias{coerce,ANY,TileDBRealizationSink-method}
\title{Write arrays to TileDB}
\description{
Write array data to a TileDB backend via \pkg{DelayedArray}'s \linkS4class{RealizationSink} machinery.
}
\section{Writing a TileDBArray}{

\code{writeTileDBArray(x, ...)} writes the matrix-like object \code{x} to a TileDB backend.
returning a \linkS4class{TileDBArray} object referring to that resource.
Arguments in \code{...} are passed to \code{TileDBRealizationSink} to configure the TileDB representation
and include \code{path}, \code{sparse}, \code{tile} and \code{ctx}.

\code{TileDBRealizationSink(dim, type="double", path=NULL, sparse=FALSE, tile=100L, ctx=NULL)}
returns a TileDBRealizationSink object that can be used to write content to a TileDB backend.
It accepts the following arguments:
\itemize{
\item \code{dim}, an integer vector (usually of length 2) to specify the array dimensions.
\item \code{type}, a string specifying the data type.
Currently only numeric, logical and integer arrays are supported.
\item \code{path}, a string containing the location of the new TileDB backend.
Defaults to a temporary directory if not specified.
\item \code{sparse} is a logical scalar indicating whether the array should be stored in sparse form.
\item \code{tile} is an integer scalar or vector specifying the tile extent for each dimension,
roughly interpreted as the size of chunks used for reading/writing the resource.
\item \code{ctx} is the TileDB context, defaulting to the output of \code{\link{tiledb_ctx}()}.
}
}

\section{Coercing to a TileDBArray}{

\code{as(x, "TileDBArray")} will coerce a matrix-like object \code{x} to a TileDBArray object.

\code{as(x, "TileDBArraySeed")} will coerce a matrix-like object \code{x} to a TileDBArraySeed object.

\code{as(x, "TileDBMatrix")} will coerce a matrix-like object \code{x} to a TileDBArraySeed object.

\code{as(x, "TileDBArray")} will coerce a TileDBRealizationSink \code{x} to a TileDBArray object.

\code{as(x, "TileDBArraySeed")} will coerce a TileDBRealizationSink \code{x} to a TileDBArraySeed object.

\code{as(x, "DelayedArray")} will coerce a TileDBRealizationSink \code{x} to a TileDBArray object.
}

\examples{
X <- matrix(rnorm(100000), ncol=200)
path <- tempfile()
out <- writeTileDBArray(X, path=path)

Y <- Matrix::rsparsematrix(1000, 1000, density=0.01)
path2 <- tempfile()
out2 <- writeTileDBArray(Y, path=path2, sparse=TRUE)

}
